Revamped consumers
